upload

+++ src/com/bridgelabz/Component/Registration.java
@@ -1,35 +1,46 @@
-//purpose :getter setter for registration
+/**
+ * purpose :getter setter for registration 
+ * Get name email phonenumber password and set values
+ */
 
 package com.bridgelabz.Component;
 
 public class Registration {
-private static String name;
-private static String email;
-private static int phoneNumber;
-private static String password;
-public static String getName() {
-	return name;
-}
-public static void setName(String name) {
-	Registration.name = name;
-}
-public static String getEmail() {
-	return email;
-}
-public static void setEmail(String email) {
-	Registration.email = email;
-}
-public int getphoneNumber() {
-	return phoneNumber;
-}
-public static void setphoneNumber(int phoneNumber) {
-	Registration.phoneNumber = phoneNumber;
-}
-public static String getPassword() {
-	return password;
-}
-public static void setPassword(String password) {
-	Registration.password = password;
-}
+	private static String name;
+	private static String email;
+	private static long phoneNumber;
+	private static String password;
+
+	public static String getName() {
+		return name;
+	}
+
+	public static void setName(String name) {
+		Registration.name = name;
+	}
+
+	public static String getEmail() {
+		return email;
+	}
+
+	public static void setEmail(String email) {
+		Registration.email = email;
+	}
+
+	public long getphoneNumber() {
+		return phoneNumber;
+	}
+
+	public static void setphoneNumber(long phoneNumber) {
+		Registration.phoneNumber = phoneNumber;
+	}
+
+	public static String getPassword() {
+		return password;
+	}
+
+	public static void setPassword(String password) {
+		Registration.password = password;
+	}
 
 }
